# [libc++] Optimize rotate